@@ -8,15 +8,15 @@
     "prettier": "prettier --write 'src/**/*.{ts,tsx,js,vue}'"
   },
   "dependencies": {
-    "@reduxjs/toolkit": "^1.8.2",
-    "clsx": "^1.1.1",
-    "evergreen-ui": "^6.9.11",
-    "framer-motion": "^6.3.11",
+    "@reduxjs/toolkit": "^1.8.3",
+    "clsx": "^1.2.1",
+    "evergreen-ui": "^6.10.3",
+    "framer-motion": "^6.5.1",
     "lodash": "^4.17.21",
-    "npm": "^8.12.1",
+    "npm": "^8.15.1",
     "phosphor-react": "^1.4.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-instantsearch-hooks-web": "^6.30.2",
     "react-redux": "^8.0.2",
     "react-router-dom": "^6.3.0",
@@ -25,15 +25,15 @@
   },
   "devDependencies": {
     "@types/lodash": "^4.14.182",
-    "@types/node": "^17.0.43",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.11",
+    "@types/node": "^18.6.2",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
     "@vitejs/plugin-react-refresh": "^1.3.6",
     "autoprefixer": "^10.4.7",
     "postcss": "^8.4.14",
-    "prettier": "2.7.0",
-    "tailwindcss": "^3.1.3",
-    "typescript": "^4.7.3",
-    "vite": "^2.9.12"
+    "prettier": "2.7.1",
+    "tailwindcss": "^3.1.6",
+    "typescript": "^4.7.4",
+    "vite": "^3.0.3"
   }
 }
